refactor(domain): split RecordTag into record/tag.rs
